feat: 호구 보고서 service 구현 #40

+++ src/main/java/com/hogu/am_i_hogu/domain/user/dto/response/HoguReportResponse.java
@@ -9,8 +9,8 @@ public record HoguReportResponse(
         String hoguLevel,
         String hoguDescription,
         List<CategoryAnalysisResponse> categoryAnalysis,
-        Long totalPostCount,
-        Long hoguPostCount,
-        Long notHoguPostCount
+        Integer totalPostCount,
+        Integer hoguPostCount,
+        Integer notHoguPostCount
 ) {
 }
